move portfolio blogs feed to content state
