+++ packages/client/src/app/pages/portfolio/blogs/blogs.component.html
@@ -1,47 +1,49 @@
-<div class="blog-feed" *ngIf="blogsData.totalDocs !== 0; else noGalleryBlogs">
-    <div class="blog-container" *ngFor="let blog of blogsData.docs | paginate: {
-        itemsPerPage: blogsData.limit,
-        currentPage: pageNum,
-        totalItems: blogsData.totalDocs
-    }">
-        <div class="blog-box">
-            <div class="blog-header">
-                <rating-icon [rating]="blog.meta.rating" [size]="'large'"></rating-icon>
-                <h1 class="header"><a [routerLink]="['/portfolio', portUser._id, portUser.username | slugify, 'blog', blog._id]">{{ blog.title }}</a></h1>
-                <div class="post-cat">
-                    <div class="info">
-                        <span>{{ blog.audit.publishedOn | localedate:'fullDate' }}</span>
-                        <span class="divider">//</span>
-                        <span [matTooltip]="'Views'" [matTooltipClass]="'offprint-tooltip'"><i-feather name="eye"></i-feather>{{ blog.stats.views }}</span>
-                        <span class="divider">//</span>
-                        <span [matTooltip]="'Comments'" [matTooltipClass]="'offprint-tooltip'"><i-feather name="message-circle"></i-feather>{{ blog.stats.comments }}</span>
-                    </div>
-                    <div class="post-tools preview">
-                        <button matRipple><i-feather name="thumbs-up"></i-feather><span class="button-text">{{ blog.stats.likes | abbreviate }}</span></button>
-                        <button matRipple class="mid"><i-feather name="thumbs-down"></i-feather><span class="button-text">{{ blog.stats.dislikes | abbreviate }}</span></button>
-                        <button matRipple class="end"><i-feather class="more" name="more-vertical"></i-feather></button>
+<ng-container *ngIf="currPageContent$ | async as blogsData">
+    <div class="blog-feed" *ngIf="blogsData.totalDocs !== 0; else noGalleryBlogs">
+        <div class="blog-container" *ngFor="let blog of blogsData.docs | paginate: {
+            itemsPerPage: blogsData.limit,
+            currentPage: pageNum,
+            totalItems: blogsData.totalDocs
+        }">
+            <div class="blog-box">
+                <div class="blog-header">
+                    <rating-icon [rating]="blog.meta.rating" [size]="'large'"></rating-icon>
+                    <h1 class="header"><a [routerLink]="['/portfolio', portUser._id, portUser.username | slugify, 'blog', blog._id]">{{ blog.title }}</a></h1>
+                    <div class="post-cat">
+                        <div class="info">
+                            <span>{{ blog.audit.publishedOn | localedate:'fullDate' }}</span>
+                            <span class="divider">//</span>
+                            <span [matTooltip]="'Views'" [matTooltipClass]="'offprint-tooltip'"><i-feather name="eye"></i-feather>{{ blog.stats.views }}</span>
+                            <span class="divider">//</span>
+                            <span [matTooltip]="'Comments'" [matTooltipClass]="'offprint-tooltip'"><i-feather name="message-circle"></i-feather>{{ blog.stats.comments }}</span>
+                        </div>
+                        <div class="post-tools preview">
+                            <button matRipple><i-feather name="thumbs-up"></i-feather><span class="button-text">{{ blog.stats.likes | abbreviate }}</span></button>
+                            <button matRipple class="mid"><i-feather name="thumbs-down"></i-feather><span class="button-text">{{ blog.stats.dislikes | abbreviate }}</span></button>
+                            <button matRipple class="end"><i-feather class="more" name="more-vertical"></i-feather></button>
+                        </div>
                     </div>
                 </div>
+                <div class="blog-content">
+                    <div class="html-description" [innerHTML]="blog.body | safeHtml"></div>
+                </div>
             </div>
-            <div class="blog-content">
-                <div class="html-description" [innerHTML]="blog.body | safeHtml"></div>
+            <div class="blog-footer">
+                <div style="text-align: center;">
+                    <a class="button" [routerLink]="['/portfolio', portUser._id, portUser.username | slugify, 'blog', blog._id]">Continue Reading</a>
+                </div>    
             </div>
+            <hr>
         </div>
-        <div class="blog-footer">
-            <div style="text-align: center;">
-                <a class="button" [routerLink]="['/portfolio', portUser._id, portUser.username | slugify, 'blog', blog._id]">Continue Reading</a>
-            </div>    
+    </div>
+    <ng-template #noGalleryBlogs>
+        <div class="empty">
+            <h3>There's nothing here...</h3>
+            <p *ngIf="currentUserIsSame(); else notSameUser">Why not head to list view and add a new blog?</p>
+            <ng-template #notSameUser>
+                <p>Check back later to see if {{ portUser.username }} adds anything!</p>
+            </ng-template>
         </div>
-        <hr>
-    </div>
-</div>
-<ng-template #noGalleryBlogs>
-    <div class="empty">
-        <h3>There's nothing here...</h3>
-        <p *ngIf="currentUserIsSame(); else notSameUser">Why not head to list view and add a new blog?</p>
-        <ng-template #notSameUser>
-            <p>Check back later to see if {{ portUser.username }} adds anything!</p>
-        </ng-template>
-    </div>
-</ng-template>
-<pagination-controls (pageChange)="onPageChange($event)" previousLabel="" nextLabel=""></pagination-controls>
+    </ng-template>
+    <pagination-controls (pageChange)="onPageChange($event)" previousLabel="" nextLabel=""></pagination-controls>
+</ng-container>
