+++ packages/client/src/app/pages/portfolio/blogs/index.ts
@@ -1,2 +1,3 @@
 export { BlogsComponent } from './blogs.component';
+export { BlogsResolver } from './blogs.resolver';
 export { BlogPageComponent } from './blog-page/blog-page.component';
