+++ packages/client/src/app/resolvers/index.ts
@@ -3,7 +3,6 @@ export { NewsFeedResolver } from './news-feed.resolver';
 export { MyWorksResolver } from './my-works.resolver';
 export { MyBlogsResolver } from './my-blogs.resolver';
 export { PortWorksResolver } from './port-works.resolver';
-export { PortBlogsResolver } from './port-blogs.resolver';
 export { MyStuffResolver } from './my-stuff.resolver';
 export { ViewContentResolver } from './view-content.resolver';
 export { SectionResolver } from './section.resolver';
